+++ twitter_bot_utils/args.py
@@ -37,19 +37,22 @@ def parent(version=None):
     return parser
 
 
-def add_logger(args):
+def add_logger(args, verbose=None, quiet=None):
     '''Set up a stdout logger'''
-    log = logging.getLogger(args.user)
+    try:
+        log = logging.getLogger(args.user)
+    except AttributeError:
+        log = logging.getLogger(args)
 
     ch = logging.StreamHandler(sys.stdout)
     ch.setFormatter(logging.Formatter('%(name)-13s: %(filename)-10s %(lineno)-3d %(message)s'))
 
-    if args.verbose:
+    if getattr(args, 'verbose', None) or verbose:
         ch.setLevel(logging.DEBUG)
     else:
         ch.setLevel(logging.INFO)
 
-    if args.quiet:
+    if getattr(args, 'quiet', None) or quiet:
         ch.setLevel(logging.ERROR)
 
     log.addHandler(ch)